refactor(detailed-view): rename the modules to DetailedGameView*
Review request: DetailedViewFilters.ts and DetailedViewModal.ts become
DetailedGameViewFilters.ts and DetailedGameViewModal.ts, with the class and
the test file renamed to match.

The element tag, page id, router name and translation keys keep the
"detailed view" wording, since that's the feature's name in the UI.
